Fix root path, xhr requests, url_for
