+++ client/.env.example
@@ -1,1 +1,2 @@
 SECRET_KEY=
+ROOT_PATH=
