+++ client/app/__init__.py
@@ -1,8 +1,13 @@
-from flask import Flask
-from config import Configs
+import os
+
 from app.routes import website_bp
+from config import Configs
+from flask import Flask
 
 app = Flask(__name__)
+app.root_path = os.getenv('ROOT_PATH')
+app.template_folder = "app/templates"
+app.static_folder = "app/static"
 app.config.from_object(Configs)
 
 app.register_blueprint(website_bp)
