+++ client/app/static/css/style.css
@@ -1,0 +1,3 @@
+body {
+    background: red;
+}
