+++ client/requirements.txt
@@ -9,5 +9,5 @@ MarkupSafe==1.1.1
 python-dotenv==0.11.0
 pytz==2019.3
 six==1.14.0
-Werkzeug==1.0.0
+Werkzeug==0.16.1
 WTForms==2.2.1
